added the functionality to check the user type before rendering components allowed for admin only

+++ src/pages/Home/index.js
@@ -10,6 +10,7 @@ import {
 import { useDispatch } from "react-redux";
 import ChangeEmployeeRoleForm from "../../components/ChangeEmployeeRoleForm";
 import DeleteEmployee from "../../components/DeleteEmployee";
+import Cookies from "js-cookie";
 function Home() {
   const [employees, setEmployees] = useState([]);
   const dispatch = useDispatch();
@@ -18,13 +19,21 @@ function Home() {
   }, []);
 
   const changeRoleHandler = (id) => {
-    dispatch(showChangeEmployeeForm());
-    dispatch(getEmployeeId(id));
+    if (isAdmin()) {
+      dispatch(showChangeEmployeeForm());
+      dispatch(getEmployeeId(id));
+    } else {
+      alert("Only Admin allowed!!");
+    }
   };
 
   const deleteEmployeeHandler = (id) => {
-    dispatch(showDeleteEmployeeConfirmation());
-    dispatch(getEmployeeId(id));
+    if (isAdmin()) {
+      dispatch(showDeleteEmployeeConfirmation());
+      dispatch(getEmployeeId(id));
+    } else {
+      alert("Only Admin allowed!!");
+    }
   };
   const getEmployees = () => {
     employeesAPI()
@@ -37,6 +46,12 @@ function Home() {
         console.log(err);
       });
   };
+
+  const isAdmin = () => {
+    // return true if the current user is admin
+    return Cookies.get("username") === "admin";
+  };
+
   return (
     <div className="container" id="employees-container">
       {employees.map((employee) => {
@@ -55,6 +70,7 @@ function Home() {
                 className="employee-Btns"
                 variant="outline-danger"
                 onClick={() => changeRoleHandler(employee.id)}
+                disabled={isAdmin() ? false : true}
               >
                 Change Role
               </Button>
@@ -64,6 +80,7 @@ function Home() {
                 className="employee-Btns"
                 variant="outline-danger"
                 onClick={() => deleteEmployeeHandler(employee.id)}
+                disabled={isAdmin() ? false : true}
               >
                 Delete Employee
               </Button>
